Fixed integer arithmetic in stdev computation

+++ src/edu/washington/escience/myriad/operator/agg/FloatAggregator.java
@@ -172,7 +172,7 @@ public final class FloatAggregator implements Aggregator {
       idx++;
     }
     if ((aggOps & AGG_OP_STDEV) != 0) {
-      double stdev = Math.sqrt(sumSquared / count - sum / count * sum / count);
+      double stdev = Math.sqrt((sumSquared / count) - ((double) (sum) / count * sum / count));
       buffer.put(idx, stdev);
       idx++;
     }
